# Optimize including schemas