--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Upcoming
 
+## 1.3.2
+
 Performance:
 
 * Optimized including other schemas in a schema, which previously caused a huge slowdown, and possibly even out-of-memory errors.