--- a/package.yaml
+++ b/package.yaml
@@ -1,5 +1,5 @@
 name: aeson-schemas
-version: 1.3.1
+version: 1.3.2
 verbatim:
   cabal-version: '>= 1.10'
 license: BSD3